// apps/frontend/src/api/generated/model/budgetPeriod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.9
 */

export type BudgetPeriod = (typeof BudgetPeriod)[keyof typeof BudgetPeriod]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BudgetPeriod = {
  monthly: 'monthly',
  weekly: 'weekly',
  biweekly: 'biweekly',
  semimonthly: 'semimonthly',
  annual: 'annual',
  custom: 'custom',
} as const